chore: add memory option, public cloud budget wording fix, current requested quota for desigin page
