--- emails/_components/Edit/quotaLookup.ts
@@ -1,37 +1,0 @@
-export const defaultCpuOptionsLookup: { [key: string]: string } = {
-  CPU_REQUEST_0_5_LIMIT_1_5: '0.5 CPU Request, 1.5 CPU Limit',
-  CPU_REQUEST_1_LIMIT_2: '1 CPU Request, 2 CPU Limit',
-  CPU_REQUEST_2_LIMIT_4: '2 CPU Request, 4 CPU Limit',
-  CPU_REQUEST_4_LIMIT_8: '4 CPU Request, 8 CPU Limit',
-  CPU_REQUEST_8_LIMIT_16: '8 CPU Request, 16 CPU Limit',
-  CPU_REQUEST_16_LIMIT_32: '16 CPU Request, 32 CPU Limit',
-  CPU_REQUEST_32_LIMIT_64: '32 CPU Request, 64 CPU Limit',
-  CPU_REQUEST_64_LIMIT_128: '64 CPU Request, 128 CPU Limit',
-};
-
-export const defaultMemoryOptionsLookup: { [key: string]: string } = {
-  MEMORY_REQUEST_2_LIMIT_4: '2 GB Request, 4 GB Limit',
-  MEMORY_REQUEST_4_LIMIT_8: '4 GB Request, 8 GB Limit',
-  MEMORY_REQUEST_8_LIMIT_16: '8 GB Request, 16 GB Limit',
-  MEMORY_REQUEST_16_LIMIT_32: '16 GB Request, 32 GB Limit',
-  MEMORY_REQUEST_32_LIMIT_64: '32 GB Request, 64 GB Limit',
-  MEMORY_REQUEST_64_LIMIT_128: '64 GB Request, 128 GB Limit',
-};
-
-export const defaultStorageOptionsLookup: { [key: string]: string } = {
-  STORAGE_1: '1 GB',
-  STORAGE_2: '2 GB',
-  STORAGE_4: '4 GB',
-  STORAGE_16: '16 GB',
-  STORAGE_32: '32 GB',
-  STORAGE_64: '64 GB',
-  STORAGE_128: '128 GB',
-  STORAGE_256: '256 GB',
-  STORAGE_512: '512 GB',
-};
-
-export const defaultProvisionedResourceValues = {
-  cpu: 'CPU: 0.5 CPU Request, 1.5 CPU Limit',
-  memory: 'Memory: 2 GB Request, 4 GB Limit',
-  storage: 'Storage: 1 GB',
-};
